har lige lagt et nyt ordenligt commit
