+++ .gitignore
@@ -1,5 +1,4 @@
 .vscode
-datasets
 __pycache__/
 notes.md
 exercises/*
@@ -9,10 +8,15 @@ include
 lib
 share
 pyvenv.cfg
-
 .DS_Store
 
-.datasets/
+code/datasets/train/*.png
+code/datasets/test/*.png
+code/datasets/train/*.m
+code/datasets/test/*.m
+code/datasets/train/*.mat
+code/datasets/test/*.mat
+*.png
 
 ## Core latex/pdflatex auxiliary files:
 *.aux
